feat: Create new CI mode
Rather than having a full-blown TUI this is a simpler flow that only logs messages
