+++ bin/hobby-installer/core/geoip.go
@@ -1,4 +1,4 @@
-package installer
+package core
 
 import (
 	"fmt"
@@ -17,23 +17,19 @@ const (
 func DownloadGeoIP() error {
 	logger := GetLogger()
 
-	// Create share directory
 	if err := os.MkdirAll(shareDir, 0755); err != nil {
 		return fmt.Errorf("failed to create share directory: %w", err)
 	}
 
-	// Check if already exists
 	if FileExists(mmdbFile) {
 		logger.WriteString("GeoIP database already exists\n")
 		return nil
 	}
 
-	// Install required tools if needed
 	if err := installGeoIPDeps(); err != nil {
 		return err
 	}
 
-	// Download and decompress
 	logger.WriteString("Downloading GeoIP database...\n")
 	cmd := exec.Command("sh", "-c",
 		fmt.Sprintf("curl -L '%s' --http1.1 | brotli --decompress --output=%s", geoIPURL, mmdbFile))
@@ -42,13 +38,11 @@ func DownloadGeoIP() error {
 	}
 	logger.WriteString("GeoIP database downloaded\n")
 
-	// Create JSON metadata file
 	jsonContent := fmt.Sprintf(`{"date": "%s"}`, time.Now().Format("2006-01-02"))
 	if err := os.WriteFile(jsonFile, []byte(jsonContent), 0644); err != nil {
 		return fmt.Errorf("failed to write GeoIP metadata: %w", err)
 	}
 
-	// Set permissions
 	os.Chmod(mmdbFile, 0644)
 	os.Chmod(jsonFile, 0644)
 
@@ -56,12 +50,10 @@ func DownloadGeoIP() error {
 }
 
 func installGeoIPDeps() error {
-	// Check if brotli is installed
 	if _, err := exec.LookPath("brotli"); err == nil {
 		return nil
 	}
 
-	// Install brotli
 	cmd := exec.Command("sudo", "apt-get", "install", "-y", "--no-install-recommends", "brotli")
 	return cmd.Run()
 }
