+++ bin/hobby-installer/core/git.go
@@ -1,4 +1,4 @@
-package installer
+package core
 
 import (
 	"fmt"
@@ -26,7 +26,6 @@ func ClonePostHog() error {
 	return err
 }
 
-// Updates the existing PostHog repository to the latest version
 func UpdatePostHog() error {
 	logger := GetLogger()
 
@@ -45,7 +44,6 @@ func UpdatePostHog() error {
 	return err
 }
 
-// CheckoutVersion checks out a specific version of PostHog
 func CheckoutVersion(version string) error {
 	if !DirExists("posthog") {
 		return fmt.Errorf("posthog directory not found")
@@ -62,12 +60,10 @@ func CheckoutVersion(version string) error {
 }
 
 func checkoutLatest() error {
-	// Fetch and reset to origin/main or current branch
 	if _, err := RunCommandWithDir("posthog", "git", "fetch", "origin"); err != nil {
 		return err
 	}
 
-	// Get current branch
 	branch, err := RunCommandWithDir("posthog", "git", "branch", "--show-current")
 	if err != nil {
 		return err
@@ -80,16 +76,13 @@ func checkoutLatest() error {
 	return err
 }
 
-// Deprecated: we don't create new `latest-release` tags anymore
 func checkoutLatestRelease() error {
 	if _, err := RunCommandWithDir("posthog", "git", "fetch", "--tags"); err != nil {
 		return err
 	}
 
-	// Get latest tag
 	out, err := RunCommandWithDir("posthog", "git", "describe", "--tags", "--abbrev=0")
 	if err != nil {
-		// Fallback: get latest tag from rev-list
 		out, err = RunCommandWithDir("posthog", "sh", "-c", "git describe --tags $(git rev-list --tags --max-count=1)")
 		if err != nil {
 			return fmt.Errorf("no release tags found")
@@ -102,7 +95,6 @@ func checkoutLatestRelease() error {
 }
 
 func checkoutSpecific(version string) error {
-	// Check if it's a commit hash (40 hex chars)
 	isCommit := regexp.MustCompile(`^[0-9a-f]{40}$`).MatchString(version)
 
 	if isCommit {
@@ -110,18 +102,15 @@ func checkoutSpecific(version string) error {
 		return err
 	}
 
-	// It's a tag/branch
 	if _, err := RunCommandWithDir("posthog", "git", "fetch", "--tags"); err != nil {
 		return err
 	}
 
-	// Try stripping "release-" prefix if present
 	releaseTag := strings.TrimPrefix(version, "release-")
 	_, err := RunCommandWithDir("posthog", "git", "checkout", releaseTag)
 	return err
 }
 
-// GetCurrentCommit returns the current commit hash
 func GetCurrentCommit() (string, error) {
 	out, err := RunCommandWithDir("posthog", "git", "rev-parse", "--short", "HEAD")
 	if err != nil {
@@ -130,17 +119,13 @@ func GetCurrentCommit() (string, error) {
 	return strings.TrimSpace(out), nil
 }
 
-// CopyComposeFiles copies docker-compose files to the working directory
 func CopyComposeFiles(version string) error {
-	// Remove existing docker-compose.yml
 	os.Remove("docker-compose.yml")
 
-	// Copy base file
 	if err := copyFile("posthog/docker-compose.base.yml", "docker-compose.base.yml"); err != nil {
 		return err
 	}
 
-	// Copy hobby file with environment variable substitution
 	return copyFileWithEnvSubst("posthog/docker-compose.hobby.yml", "docker-compose.yml", version)
 }
 
@@ -152,7 +137,6 @@ func copyFileWithEnvSubst(src, dst, version string) error {
 
 	content := string(data)
 
-	// Get registry URL from env or .env, default to posthog/posthog
 	registryURL := os.Getenv("REGISTRY_URL")
 	if registryURL == "" {
 		registryURL = ReadEnvValue("REGISTRY_URL")
@@ -161,12 +145,10 @@ func copyFileWithEnvSubst(src, dst, version string) error {
 		registryURL = "posthog/posthog"
 	}
 
-	// Use the version selected by the user
 	if version == "" {
 		version = "latest"
 	}
 
-	// Replace ${VAR} and $VAR patterns
 	content = strings.ReplaceAll(content, "${REGISTRY_URL}", registryURL)
 	content = strings.ReplaceAll(content, "$REGISTRY_URL", registryURL)
 	content = strings.ReplaceAll(content, "${POSTHOG_APP_TAG}", version)
@@ -183,13 +165,11 @@ func copyFile(src, dst string) error {
 	return os.WriteFile(dst, data, 0644)
 }
 
-// CreateComposeScripts creates the compose/start and compose/wait scripts
 func CreateComposeScripts() error {
 	if err := os.MkdirAll("compose", 0755); err != nil {
 		return err
 	}
 
-	// Create start script
 	startScript := `#!/bin/bash
 ./compose/wait
 ./bin/migrate
@@ -199,7 +179,6 @@ func CreateComposeScripts() error {
 		return err
 	}
 
-	// Create temporal-django-worker script
 	temporalScript := `#!/bin/bash
 ./bin/temporal-django-worker
 `
@@ -207,7 +186,6 @@ func CreateComposeScripts() error {
 		return err
 	}
 
-	// Create wait script
 	waitScript := `#!/usr/bin/env python3
 
 import socket
@@ -231,12 +209,10 @@ loop()
 	return os.WriteFile("compose/wait", []byte(waitScript), 0755)
 }
 
-// SetupGit installs git if needed
 func SetupGit() error {
 	if _, err := exec.LookPath("git"); err == nil {
-		return nil // Already installed
+		return nil
 	}
-	// Install git
 	cmd := exec.Command("sudo", "apt", "install", "-y", "git")
 	return cmd.Run()
 }
