+++ bin/hobby-installer/core/upgrade.go
@@ -1,15 +1,12 @@
-package installer
+package core
 
 import (
-	"bufio"
 	"fmt"
 	"os"
-	"regexp"
 	"strings"
 	"time"
 )
 
-// UpgradeCheck holds the result of pre-upgrade checks
 type UpgradeCheck struct {
 	HasNamedVolumes        bool
 	NeedsPostgresMigration bool
@@ -21,21 +18,17 @@ type UpgradeCheck struct {
 func CheckUpgradeRequirements() (*UpgradeCheck, error) {
 	check := &UpgradeCheck{}
 
-	// Check for named volumes
 	postgres, clickhouse := CheckDockerVolumes()
 	check.HasNamedVolumes = postgres && clickhouse
 
-	// Check for postgres 12 in compose file
 	check.NeedsPostgresMigration = checkPostgres12InCompose()
 
-	// Check storage migration status
 	check.StorageMigrationStatus = ReadEnvValue("SESSION_RECORDING_STORAGE_MIGRATED_TO_SEAWEEDFS")
 	check.NeedsStorageMigration = check.StorageMigrationStatus == ""
 
 	return check, nil
 }
 
-// Returns warning message if named volumes are missing (pre-1.39 installation)
 func GetVolumeWarning() string {
 	postgres, clickhouse := CheckDockerVolumes()
 	if postgres && clickhouse {
@@ -75,7 +68,6 @@ func BackupPostgres12() (string, error) {
 		return "", fmt.Errorf("backup failed: %w", err)
 	}
 
-	// Verify backup
 	if !FileExists(backupFile) {
 		return "", fmt.Errorf("backup file was not created")
 	}
@@ -105,96 +97,19 @@ func MigratePostgres12To15(backupFile string) error {
 }
 
 func RestorePostgres15(backupFile string) error {
-	// Start just the DB
 	if err := DockerComposeUpDB(); err != nil {
 		return fmt.Errorf("failed to start database: %w", err)
 	}
 
-	// Wait for DB to be ready
 	time.Sleep(20 * time.Second)
 
-	// Restore backup
 	if err := RestorePostgres(backupFile); err != nil {
 		return fmt.Errorf("failed to restore backup: %w", err)
 	}
 
-	// Upgrade password to SCRAM-SHA-256
 	cmd, args := GetDockerComposeCommand()
 	fullArgs := append(args, "exec", "-T", "db", "psql", "-U", "posthog", "-c", "ALTER USER posthog WITH PASSWORD 'posthog';")
 	RunCommand(cmd, fullArgs...)
-
-	return nil
-}
-
-func FixEnvQuoting() error {
-	data, err := os.ReadFile(".env")
-	if err != nil {
-		return nil // File doesn't exist, nothing to fix
-	}
-
-	var result strings.Builder
-	scanner := bufio.NewScanner(strings.NewReader(string(data)))
-
-	for scanner.Scan() {
-		line := scanner.Text()
-
-		// Skip empty lines and comments
-		if strings.TrimSpace(line) == "" || strings.HasPrefix(strings.TrimSpace(line), "#") {
-			result.WriteString(line)
-			result.WriteString("\n")
-			continue
-		}
-
-		// Split on first =
-		idx := strings.Index(line, "=")
-		if idx == -1 {
-			result.WriteString(line)
-			result.WriteString("\n")
-			continue
-		}
-
-		key := strings.TrimSpace(line[:idx])
-		value := strings.TrimSpace(line[idx+1:])
-
-		// Already quoted?
-		if (strings.HasPrefix(value, "\"") && strings.HasSuffix(value, "\"")) ||
-			(strings.HasPrefix(value, "'") && strings.HasSuffix(value, "'")) {
-			result.WriteString(line)
-			result.WriteString("\n")
-			continue
-		}
-
-		// Quote if contains space, comma, or ://
-		needsQuote := strings.ContainsAny(value, " ,") || strings.Contains(value, "://")
-		if needsQuote {
-			value = strings.ReplaceAll(value, "\"", "\\\"")
-			result.WriteString(fmt.Sprintf("%s=\"%s\"\n", key, value))
-		} else {
-			result.WriteString(line)
-			result.WriteString("\n")
-		}
-	}
-
-	return os.WriteFile(".env", []byte(result.String()), 0600)
-}
-
-// Validates the .env file has required keys
-func ValidateEnvForUpgrade() error {
-	required := []string{"POSTHOG_SECRET", "DOMAIN"}
-	for _, key := range required {
-		if ReadEnvValue(key) == "" {
-			return fmt.Errorf("missing required env var: %s", key)
-		}
-	}
-
-	// Validate encryption key format if present
-	encKey := ReadEnvValue("ENCRYPTION_SALT_KEYS")
-	if encKey != "" {
-		// Should be 32 hex chars
-		if !regexp.MustCompile(`^[A-Za-z0-9_-]{32}$`).MatchString(encKey) {
-			return fmt.Errorf("ENCRYPTION_SALT_KEYS is not in correct format")
-		}
-	}
 
 	return nil
 }
