--- bin/hobby-installer/steps/install.go
@@ -1,348 +1,0 @@
-package steps
-
-import (
-	"fmt"
-	"time"
-
-	"github.com/charmbracelet/bubbles/spinner"
-	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
-	"github.com/posthog/posthog/bin/hobby-installer/installer"
-	"github.com/posthog/posthog/bin/hobby-installer/ui"
-)
-
-type installStep int
-
-const (
-	installStepTelemetryStart installStep = iota
-	installStepGit
-	installStepClone
-	installStepCheckout
-	installStepEnv
-	installStepGeoIP
-	installStepScripts
-	installStepCopyCompose
-	installStepDockerSetup
-	installStepPull
-	installStepAsyncMigrations
-	installStepStart
-	installStepHealth
-	installStepTelemetryComplete
-	installStepDone
-)
-
-type installStatus int
-
-const (
-	installPending installStatus = iota
-	installRunning
-	installSuccess
-	installFailed
-	installSkipped
-)
-
-type installItem struct {
-	name   string
-	status installStatus
-	detail string
-}
-
-type InstallModel struct {
-	steps       []installItem
-	currentStep installStep
-	spinner     spinner.Model
-	isUpgrade   bool
-	version     string
-	domain      string
-	err         error
-	width       int
-	height      int
-}
-
-type stepResultMsg struct {
-	step   installStep
-	err    error
-	detail string
-}
-
-func NewInstallModel() InstallModel {
-	s := spinner.New()
-	s.Spinner = spinner.Dot
-	s.Style = ui.SpinnerStyle
-
-	return InstallModel{
-		steps: []installItem{
-			{name: "Send telemetry", status: installPending},
-			{name: "Setup git", status: installPending},
-			{name: "Clone/update PostHog repository", status: installPending},
-			{name: "Checkout version", status: installPending},
-			{name: "Generate configuration", status: installPending},
-			{name: "Download GeoIP database", status: installPending},
-			{name: "Create startup scripts", status: installPending},
-			{name: "Copy Docker Compose files", status: installPending},
-			{name: "Setup Docker", status: installPending},
-			{name: "Pull Docker images", status: installPending},
-			{name: "Check async migrations", status: installPending},
-			{name: "Start PostHog stack", status: installPending},
-			{name: "Wait for PostHog to be ready", status: installPending},
-			{name: "Send completion telemetry", status: installPending},
-		},
-		currentStep: installStepTelemetryStart,
-		spinner:     s,
-	}
-}
-
-func (m *InstallModel) SetConfig(isUpgrade bool, version, domain string) {
-	m.isUpgrade = isUpgrade
-	m.version = version
-	m.domain = domain
-}
-
-func (m InstallModel) Init() tea.Cmd {
-	return tea.Batch(
-		m.spinner.Tick,
-		m.runStep(installStepTelemetryStart),
-	)
-}
-
-func (m InstallModel) runStep(step installStep) tea.Cmd {
-	return func() tea.Msg {
-		var err error
-		var detail string
-
-		switch step {
-		case installStepTelemetryStart:
-			installer.SendInstallStartEvent(m.domain)
-			detail = "sent"
-
-		case installStepGit:
-			// Run apt update before installing git
-			_ = installer.AptUpdate()
-			err = installer.SetupGit()
-			detail = "git ready"
-
-		case installStepClone:
-			// Auto-detect: if posthog dir exists, update; otherwise clone
-			if installer.DirExists("posthog") {
-				err = installer.UpdatePostHog()
-				detail = "updated"
-			} else {
-				err = installer.ClonePostHog()
-				detail = "cloned"
-			}
-
-		case installStepCheckout:
-			err = installer.CheckoutVersion(m.version)
-			if err == nil {
-				commit, _ := installer.GetCurrentCommit()
-				detail = fmt.Sprintf("at %s", commit)
-			}
-
-		case installStepEnv:
-			// Auto-detect: if .env exists, update; otherwise create
-			if installer.FileExists(".env") {
-				// Fix quoting issues from previous versions
-				if fixErr := installer.FixEnvQuoting(); fixErr != nil {
-					installer.GetLogger().WriteString(fmt.Sprintf("Warning: could not fix .env quoting: %v\n", fixErr))
-				}
-				err = installer.UpdateEnvForUpgrade(m.version)
-				detail = "updated"
-			} else {
-				config, configErr := installer.NewEnvConfig(m.domain, m.version)
-				if configErr != nil {
-					err = configErr
-				} else {
-					err = config.WriteEnvFile()
-					detail = "created .env"
-				}
-			}
-
-		case installStepGeoIP:
-			if installer.GeoIPExists() {
-				detail = "already exists"
-			} else {
-				err = installer.DownloadGeoIP()
-				detail = "downloaded"
-			}
-
-		case installStepScripts:
-			err = installer.CreateComposeScripts()
-			detail = "created"
-
-		case installStepCopyCompose:
-			err = installer.CopyComposeFiles(m.version)
-			detail = "copied"
-
-		case installStepDockerSetup:
-			if installer.IsDockerInstalled() && installer.IsDockerRunning() {
-				detail = "Docker ready"
-			} else if !installer.IsDockerInstalled() {
-				err = installer.InstallDocker()
-				if err == nil {
-					err = installer.InstallDockerCompose()
-				}
-				detail = "installed"
-			} else {
-				//nolint:revive,stylecheck // ST1005: error strings should not be capitalized
-				err = fmt.Errorf("docker installed but not running")
-			}
-
-		case installStepPull:
-			err = installer.DockerComposePull()
-			detail = "images pulled"
-
-		case installStepAsyncMigrations:
-			// Only run async migrations check if this is an upgrade (volumes exist)
-			hasPostgres, hasClickhouse := installer.CheckDockerVolumes()
-			if hasPostgres || hasClickhouse {
-				err = installer.RunAsyncMigrationsCheck()
-				detail = "checked"
-			} else {
-				detail = "skipped (new install)"
-			}
-
-		case installStepStart:
-			_ = installer.DockerComposeStop() // Stop any existing stack
-			err = installer.DockerComposeUpWithRetry(3)
-			detail = "started"
-
-		case installStepHealth:
-			err = installer.WaitForHealth(10 * time.Minute)
-			if err == nil {
-				detail = "PostHog is up!"
-			}
-
-		case installStepTelemetryComplete:
-			installer.SendInstallCompleteEvent(m.domain)
-			detail = "sent"
-		}
-
-		return stepResultMsg{step: step, err: err, detail: detail}
-	}
-}
-
-func (m InstallModel) Update(msg tea.Msg) (InstallModel, tea.Cmd) {
-	switch msg := msg.(type) {
-	case tea.WindowSizeMsg:
-		m.width = msg.Width
-		m.height = msg.Height
-		return m, nil
-
-	case spinner.TickMsg:
-		var cmd tea.Cmd
-		m.spinner, cmd = m.spinner.Update(msg)
-		return m, cmd
-
-	case stepResultMsg:
-		idx := int(msg.step)
-		if msg.err != nil {
-			m.steps[idx].status = installFailed
-			m.steps[idx].detail = msg.err.Error()
-			m.err = msg.err
-			return m, func() tea.Msg {
-				return ErrorMsg{Err: msg.err}
-			}
-		}
-
-		m.steps[idx].status = installSuccess
-		m.steps[idx].detail = msg.detail
-
-		// Move to next step
-		nextStep := msg.step + 1
-		if nextStep >= installStepDone {
-			m.currentStep = installStepDone
-			return m, func() tea.Msg {
-				return StepCompleteMsg{Data: nil}
-			}
-		}
-
-		m.currentStep = nextStep
-		m.steps[int(nextStep)].status = installRunning
-
-		// Skip async migrations for new installs (no existing volumes)
-		if nextStep == installStepAsyncMigrations {
-			hasPostgres, hasClickhouse := installer.CheckDockerVolumes()
-			if !hasPostgres && !hasClickhouse {
-				m.steps[int(nextStep)].status = installSkipped
-				m.steps[int(nextStep)].detail = "new install"
-				nextStep++
-				if nextStep < installStepDone {
-					m.currentStep = nextStep
-					m.steps[int(nextStep)].status = installRunning
-				}
-			}
-		}
-
-		return m, m.runStep(nextStep)
-	}
-
-	return m, nil
-}
-
-func (m InstallModel) View() string {
-	var lines []string
-
-	for i, step := range m.steps {
-		var icon string
-		var style lipgloss.Style
-
-		switch step.status {
-		case installPending:
-			icon = "○"
-			style = ui.MutedStyle
-		case installRunning:
-			icon = m.spinner.View()
-			style = lipgloss.NewStyle().Foreground(ui.ColorPrimary)
-		case installSuccess:
-			icon = ui.Checkmark()
-			style = ui.SuccessStyle
-		case installFailed:
-			icon = ui.Cross()
-			style = ui.ErrorStyle
-		case installSkipped:
-			icon = "◌"
-			style = ui.MutedStyle
-		}
-
-		line := fmt.Sprintf("  %s %s", icon, style.Render(step.name))
-		if step.detail != "" && step.status != installPending && step.status != installRunning {
-			line += ui.MutedStyle.Render(fmt.Sprintf(" (%s)", step.detail))
-		}
-
-		// Set first step as running if it's pending
-		if i == 0 && step.status == installPending && m.currentStep == 0 {
-			m.steps[0].status = installRunning
-		}
-
-		lines = append(lines, line)
-	}
-
-	title := "Setting up PostHog"
-
-	var footer string
-	if m.currentStep == installStepDone {
-		footer = "\n" + ui.SuccessStyle.Render("✓ Installation complete!")
-	} else if m.err != nil {
-		footer = "\n" + ui.ErrorStyle.Render("✗ Installation failed: "+m.err.Error())
-	}
-
-	// Always show log panel
-	logLines := installer.GetLogger().GetLines(5)
-	logPanel := ui.RenderLogPanel(logLines, 100, 7)
-
-	content := lipgloss.JoinVertical(
-		lipgloss.Left,
-		ui.TitleStyle.Render(title),
-		"",
-		ui.SubtitleStyle.Render(fmt.Sprintf("Version: %s | Domain: %s", m.version, m.domain)),
-		"",
-		lipgloss.JoinVertical(lipgloss.Left, lines...),
-		footer,
-		"",
-		logPanel,
-	)
-
-	return lipgloss.NewStyle().
-		Padding(2, 4).
-		Render(content)
-}
