+++ bin/hobby-installer/tui/steps/complete.go
