+++ bin/hobby-installer/tui/steps/messages.go
@@ -9,4 +9,3 @@ type StepCompleteMsg struct {
 type ErrorMsg struct {
 	Err error
 }
-
